+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.7.3] - 2025-01-25
+
 ### Changed
 
 - Make Git plugin optional - if not installed, variables related to Git will not be available
@@ -234,7 +236,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release
 
-[Unreleased]: https://github.com/pandier/intellij-discord-rp/compare/v1.7.2...HEAD
+[Unreleased]: https://github.com/pandier/intellij-discord-rp/compare/v1.7.3...HEAD
+[1.7.3]: https://github.com/pandier/intellij-discord-rp/compare/v1.7.2...v1.7.3
 [1.7.2]: https://github.com/pandier/intellij-discord-rp/compare/v1.7.1...v1.7.2
 [1.7.1]: https://github.com/pandier/intellij-discord-rp/compare/v1.7.0...v1.7.1
 [1.7.0]: https://github.com/pandier/intellij-discord-rp/compare/v1.6.2...v1.7.0